refactor: restructure recipe model and update router to use ctrlWrapper
