+++ src/routers/recipes.js
@@ -1,8 +1,9 @@
 import express from 'express';
 import { getRecipeByIdController } from '../controllers/recipes';
+import { ctrlWrapper } from '../utils/ctrlWrapper.js';
 
 const router = express.Router();
 
-router.get('/:id', getRecipeByIdController);
+router.get('/:id', ctrlWrapper(getRecipeByIdController));
 
 export default router;
