+++ src/services/recipes.js
@@ -1,4 +1,4 @@
-import Recipe from '../models/Recipe.js';
+import { Recipe } from '../db/models/recipe.js';
 
 export const getRecipeById = async (id) => {
   return await Recipe.findById(id);
